weather/requirements.txt
Django
django-filter
djangorestframework
pandas
requests
